Rename misspelled directory, add margin to testimonial section
